+++ Program.cs
@@ -1,23 +1,14 @@
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Net.Mail;
-using System.Net.Sockets;
-using System.Text.Json;
-using System.Web;
-using UtilityBelt.Helpers;
-using UtilityBelt.Models;
-using Microsoft.Extensions.Logging;
-using UtilityBelt.Logging;
 using System.Composition;
 using System.Composition.Hosting;
+using System.Linq;
 using System.Reflection;
+using UtilityBelt.Logging;
+using UtilityBelt.Models;
 
 namespace UtilityBelt
 {
@@ -188,365 +179,6 @@ namespace UtilityBelt
     }
 
     #endregion Choice Handler
-
-    #region Choice Processors
-
-    #region Random quote
-
-    private static void RandomQuoteGarden()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomQuoteGarden)}");
-      IWebProxy defaultWebProxy = WebRequest.DefaultWebProxy;
-      defaultWebProxy.Credentials = CredentialCache.DefaultCredentials;
-
-      string content = string.Empty;
-      string randomQuoteUrl = "https://quote-garden.herokuapp.com/api/v2/quotes/random";
-      using (var wc = new WebClient() { Proxy = defaultWebProxy })
-      {
-        content = wc.DownloadString(randomQuoteUrl);
-      }
-
-      RandomQuoteModel quote = JsonSerializer.Deserialize<RandomQuoteModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(@$"Quote -- {quote.Quote.Text}");
-      Console.WriteLine(@$"From -- {quote.Quote.Author}");
-
-      Console.WriteLine();
-    }
-
-    #endregion Random quote
-
-    #region Gender from name
-
-    private static void GenderFromName()
-    {
-      logger.LogInformation($"User Choice : {nameof(GenderFromName)}");
-      Console.Write("Type the name, and then press Enter: ");
-      string enteredName = Console.ReadLine().ToLower();
-
-      if (string.IsNullOrEmpty(enteredName))
-      {
-        Console.WriteLine("No name provided!");
-        Console.WriteLine();
-        return;
-      }
-
-      IWebProxy defaultWebProxy = WebRequest.DefaultWebProxy;
-      defaultWebProxy.Credentials = CredentialCache.DefaultCredentials;
-
-      string content = string.Empty;
-      string genderizatorUrl = $"https://api.genderize.io?name={enteredName}";
-      using (var wc = new WebClient() { Proxy = defaultWebProxy })
-      {
-        content = wc.DownloadString(genderizatorUrl);
-      }
-
-      GenderizatorModel genderResult = JsonSerializer.Deserialize<GenderizatorModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(@$"The gender is {genderResult.Gender} with a probability of {genderResult.Probability}");
-
-      Console.WriteLine();
-    }
-
-    #endregion Gender from name
-
-    #region Dad Joke
-
-    private static void DadJoke()
-    {
-      logger.LogInformation($"User Choice : {nameof(DadJoke)}");
-      WebRequest request = WebRequest.Create("https://icanhazdadjoke.com");
-
-      request.Headers.Add("User-Agent", "GitHub library https://github.com/aherd2985/UtilityBelt");
-      request.Headers.Add("Accept", "application/json");
-
-      // Get the response.
-      WebResponse response = request.GetResponse();
-
-      var status = ((HttpWebResponse)response).StatusCode;
-
-      if (status != HttpStatusCode.OK)
-      {
-        Console.WriteLine(@"There was an error retrieving the joke. Please try again later.");
-      }
-      else
-      {
-        using Stream dataStream = response.GetResponseStream();
-
-        if (dataStream != null)
-        {
-          StreamReader reader = new StreamReader(dataStream);
-          // Read the content.
-          string responseFromServer = reader.ReadToEnd();
-          // Display the content.
-          DadJokeModel dadJoke = JsonSerializer.Deserialize<DadJokeModel>(responseFromServer);
-          Console.WriteLine(@"Random Dad Joke:");
-          Console.WriteLine(dadJoke.Joke);
-        }
-        else
-        {
-          Console.WriteLine(@"The joke data is null. This is probably bad.");
-        }
-      }
-
-      // Close the response.
-      response.Close();
-    }
-
-    #endregion Dad Joke
-
-    #region Random pada fact
-
-    private static void RandomPandaFact()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomPandaFact)}");
-      string content;
-      string url = "https://some-random-api.ml/facts/panda";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(url);
-      }
-
-      PandaFactModel pandaFact = JsonSerializer.Deserialize<PandaFactModel>(content);
-      Console.WriteLine();
-      Console.ForegroundColor = ConsoleColor.Yellow;
-      Console.WriteLine(pandaFact.Fact);
-      Console.WriteLine();
-    }
-
-    #endregion Random pada fact
-
-    #region Random fox fact
-
-    private static void RandomFoxFact()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomFoxFact)}");
-      string content;
-      string url = "https://some-random-api.ml/facts/fox";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(url);
-      }
-
-      PandaFactModel foxFact = JsonSerializer.Deserialize<PandaFactModel>(content);
-      Console.WriteLine();
-      Console.ForegroundColor = ConsoleColor.Yellow;
-      Console.WriteLine(foxFact.Fact);
-      Console.WriteLine();
-    }
-
-    #endregion Random fox fact
-
-    #region HostToIp
-
-    private static void HostToIp()
-    {
-      while (true)
-      {
-        logger.LogInformation($"User Choice : {nameof(HostToIp)}");
-        Console.Write("Please enter a hostname: ");
-        var hostname = Console.ReadLine();
-        if (string.IsNullOrEmpty(hostname) && Uri.CheckHostName(hostname) != UriHostNameType.Unknown)
-        {
-          continue;
-        }
-
-        Console.WriteLine($"Hostname: {hostname}");
-        try
-        {
-          var ipAddresses = Dns.GetHostAddresses(hostname);
-          for (var i = 0; i < ipAddresses.Length; i++)
-          {
-            Console.WriteLine($"IP[{i + 1}]: {ipAddresses[i]}");
-          }
-
-          break;
-        }
-        catch (SocketException e)
-        {
-          Console.WriteLine("Invalid hostname - " + e.Message);
-          break;
-        }
-        catch (Exception e)
-        {
-          Console.WriteLine("An error occurred: " + e.Message);
-        }
-      }
-    }
-
-    #endregion HostToIp
-
-    #region Ghost Game
-
-    private static void GhostGame()
-    {
-      logger.LogInformation($"User Choice : {nameof(GhostGame)}");
-      Random rNumber = new Random();
-      int score = 0;
-      int randomNum = rNumber.Next(1, 4);
-      int numberInput;
-      bool validDoor = false;
-      bool validNumber;
-
-      while (true)
-      {
-        Console.WriteLine("\nThere are three doors ahead! \n");
-        Console.WriteLine("A ghost is behind one of them :O! \n");
-        Console.WriteLine("Which door do you open?!?! \n");
-        Console.WriteLine("1, 2 or 3?\n");
-
-        do
-        {
-          validNumber = int.TryParse(Console.ReadLine(), out numberInput);
-
-          if (Enumerable.Range(1, 3).Contains(numberInput) && validNumber)//If user selected a valid door ...
-          {
-            validDoor = true;
-          }
-          else
-          {
-            Console.WriteLine("");
-            if (validNumber)
-            {
-              Console.WriteLine("\nThere are only three doors ahead. \n");
-            }
-            else
-            {
-              Console.WriteLine("Input was not an integer.");
-            }
-            Console.WriteLine("Please select door 1, 2, or 3.\n");
-            validDoor = false;
-          }
-        } while (validDoor == false);
-
-        if (numberInput == randomNum)
-        {
-          Console.WriteLine("\nBoo!! GHOST!");
-          break;
-        }
-        else
-        {
-          Console.WriteLine("\nNo ghost, Phew!");
-          Console.WriteLine("\nPress any key to enter the next room!");
-          Console.ReadKey();
-          randomNum = rNumber.Next(1, 4);
-          score += 1;
-        }
-      }
-      Console.WriteLine("Run!!!!!");
-      Console.WriteLine("\nGame Over! Score: " + score);
-    }
-
-    #endregion Ghost Game
-
-    #region Breaking Bad Quotes
-
-    private static void BreakingBadQuotes()
-    {
-      logger.LogInformation($"User Choice : {nameof(BreakingBadQuotes)}");
-      IWebProxy defaultWebProxy = WebRequest.DefaultWebProxy;
-      defaultWebProxy.Credentials = CredentialCache.DefaultCredentials;
-
-      string content = string.Empty;
-      string breakingBadQuoteUrl = $"https://breaking-bad-quotes.herokuapp.com/v1/quotes";
-      using (var wc = new WebClient() { Proxy = defaultWebProxy })
-      {
-        content = wc.DownloadString(breakingBadQuoteUrl);
-      }
-
-      List<BreakingBadQuoteModel> quote = JsonSerializer.Deserialize<List<BreakingBadQuoteModel>>(content);
-      Console.WriteLine();
-      Console.WriteLine(@$"Quote -- {quote.FirstOrDefault().Quote}");
-      Console.WriteLine(@$"From -- {quote.FirstOrDefault().Author}");
-
-      Console.WriteLine();
-    }
-
-    #endregion Breaking Bad Quotes
-
-    #region Digital Ocean
-
-    private static void DigitalOceanStatus()
-    {
-      logger.LogInformation($"User Choice : {nameof(DigitalOceanStatus)}");
-      IWebProxy defaultWebProxy = WebRequest.DefaultWebProxy;
-      defaultWebProxy.Credentials = CredentialCache.DefaultCredentials;
-
-      string content = string.Empty;
-      string digitalOceanStatusUrl = $"https://s2k7tnzlhrpw.statuspage.io/api/v2/status.json";
-      using (var wc = new WebClient() { Proxy = defaultWebProxy })
-      {
-        content = wc.DownloadString(digitalOceanStatusUrl);
-      }
-
-      DigitalOceanStatusModel digitalOceanStatus = JsonSerializer.Deserialize<DigitalOceanStatusModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(@$"Indicator -- {digitalOceanStatus.Status.Indicator}");
-      Console.WriteLine(@$"Description -- {digitalOceanStatus.Status.Description}");
-
-      Console.WriteLine();
-    }
-
-    #endregion Digital Ocean
-
-    #region Random User Generator
-
-    private static void RandomUserGenerator()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomUserGenerator)}");
-      string content = string.Empty;
-      string randomUserUrl = @"https://randomuser.me/api";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(randomUserUrl);
-      }
-
-      RandomUser randomUser = JsonSerializer.Deserialize<RandomUser>(content);
-      Console.WriteLine();
-      Console.ForegroundColor = ConsoleColor.Cyan;
-
-      int cnt = 1;
-
-      foreach (var user in randomUser.results)
-      {
-        Console.WriteLine($"User Generated Nbr: {cnt}");
-        Console.WriteLine($"Gender: {user.Gender}");
-        Console.WriteLine($"Name: {user.Name.First} {user.Name.Last}");
-        Console.WriteLine($"Email: {user.Email}");
-        Console.WriteLine($"Phone: {user.Phone}");
-        Console.WriteLine("");
-      }
-
-      Console.WriteLine();
-    }
-
-    #endregion Random User Generator
-
-    #region Random Programmer Quote
-
-    private static void RandomProgrammerQuote()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomProgrammerQuote)}");
-      IWebProxy defaultWebProxy = WebRequest.DefaultWebProxy;
-      defaultWebProxy.Credentials = CredentialCache.DefaultCredentials;
-
-      string content = string.Empty;
-      string programmerQuoteUrl = $"https://programming-quotes-api.herokuapp.com/quotes/random";
-      using (var wc = new WebClient() { Proxy = defaultWebProxy })
-      {
-        content = wc.DownloadString(programmerQuoteUrl);
-      }
-
-      ProgrammingQuoteModel programmingQuote = JsonSerializer.Deserialize<ProgrammingQuoteModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(@$"{programmingQuote.Quote}");
-      Console.WriteLine(@$"by {programmingQuote.Author}");
-      Console.WriteLine();
-    }
-
-    #endregion Random Programmer Quote
-
-    #endregion Choice Processors
 
     #region Utility
